refactor(filters): extract pipeline definition and functions to separate file
